pyproject.toml
[tool.poetry]
name = "python-ai"
version = "0.1.0"
description = ""
authors = ["Musawir Hussain <musawir.hussain@outgive.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.1"
flask = "^3.1.0"
langchain = "^0.3.9"
faiss-cpu = "^1.9.0.post1"
beautifulsoup4 = "^4.12.3"
sentence-transformers = "^3.3.1"
langchain-community = "^0.3.8"
tiktoken = "^0.8.0"
langchain-openai = "^0.2.10"
onnxruntime= "1.14.1"
chromadb = "^0.5.21"
torch = "^2.5.1"
openai = "^1.56.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
